completed user maintenance and user access/roles screen

+++ src/panel_manageusers.java
@@ -8,10 +8,12 @@ import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Vector;
 
 public class panel_manageusers extends JPanel {
     JLabel lbl_title, lbl_currentusers;
+    public static JLabel lbl_status;
     JTable tbl_currentusers;
     JButton btn_add, btn_modify, btn_delete;
     Vector<String> headers = new Vector<String>();
@@ -21,7 +23,7 @@ public class panel_manageusers extends JPanel {
     Vector<Vector<Object>> data;
     Vector<Object> record;
     ListSelectionModel listselectionmodel;
-    String buff_userid;
+    String buff_userid, buff_username, buff_fullname, buff_password, buff_accounttype;
 
     public panel_manageusers() throws Exception {
         setLayout(new GridBagLayout());
@@ -32,6 +34,7 @@ public class panel_manageusers extends JPanel {
 
         lbl_title = new JLabel("User Maintenance");
         lbl_currentusers = new JLabel("Current Users");
+        lbl_status = new JLabel();
 
         btn_add = new JButton("Add User");
         btn_modify = new JButton("Modify User");
@@ -109,6 +112,13 @@ public class panel_manageusers extends JPanel {
 
         c.gridx = 2;
         add(btn_delete, c);
+
+        //5th row
+        c.gridx = 0;
+        c.gridy = 4;
+        c.insets = new Insets(25,0,0,10);
+        add(lbl_status, c);
+
     }
 
     public void refreshTable(ResultSet rset1) throws Exception
@@ -173,36 +183,75 @@ public class panel_manageusers extends JPanel {
             }
 
             if (source == btn_modify) {
+                dialog_updateuser si = null;
+                try {
+                    si = new dialog_updateuser(buff_username, buff_fullname, buff_password, buff_accounttype);
+                } catch (Exception exception) {
+                    exception.printStackTrace();
+                }
+                si.setSize(new Dimension(450,350));
+                si.addWindowListener(new WindowAdapter() {
+                    @Override
+                    public void windowClosed(WindowEvent e) {
+                        dbconnect conn2 = new dbconnect();
 
+                        try
+                        {
+                            rset = conn2.getUserlist();
+                            refreshTable(rset);
+                            conn2.close();
+                        }
+                        catch (Exception j)
+                        {
+                            System.out.println(j.getMessage());
+                        }
+                    }
+                });
 
+                si.setLocationRelativeTo(null);
+                si.setModal(true);
+                si.setVisible(true);
+
+                btn_modify.setEnabled(false);
+                btn_delete.setEnabled(false);
             }
 
             if (source == btn_delete)
             {
-                dbconnect conn3 = new dbconnect();
+                int response = JOptionPane.showConfirmDialog(null, "Do you really want to delete user " + buff_username + "?", "Delete User", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
 
-                try {
-                    conn3.deleteUser(buff_userid);
-                    conn3.close();
-                } catch (Exception exception) {
-                    exception.printStackTrace();
+                if (response == JOptionPane.NO_OPTION) {
+
+                } else if (response == JOptionPane.YES_OPTION) {
+                    dbconnect conn3 = new dbconnect();
+
+                    try {
+                        conn3.deleteUser(buff_userid);
+                        conn3.close();
+                    } catch (Exception exception) {
+                        exception.printStackTrace();
+                    }
+
+                    dbconnect conn2 = new dbconnect();
+
+                    try
+                    {
+                        rset = conn2.getUserlist();
+                        refreshTable(rset);
+                        conn2.close();
+                    }
+                    catch (Exception j)
+                    {
+                        System.out.println(j.getMessage());
+                    }
+
+                    btn_delete.setEnabled(false);
+                    btn_modify.setEnabled(false);
+
+                    lbl_status.setText("Deleted user " + buff_username + " from database!");
+                } else if (response == JOptionPane.CLOSED_OPTION) {
+                    System.out.println("Closed window");
                 }
-
-                dbconnect conn2 = new dbconnect();
-
-                try
-                {
-                    rset = conn2.getUserlist();
-                    refreshTable(rset);
-                    conn2.close();
-                }
-                catch (Exception j)
-                {
-                    System.out.println(j.getMessage());
-                }
-
-                btn_delete.setEnabled(false);
-                btn_modify.setEnabled(false);
             }
         }
     }
@@ -222,10 +271,42 @@ public class panel_manageusers extends JPanel {
                 int minIndex = lsm.getMinSelectionIndex();
                 int maxIndex = lsm.getMaxSelectionIndex();
 
-                Object x;
+                Object x, y, z, w;
 
                 x = tbl_currentusers.getValueAt(maxIndex, 0);
                 buff_userid = x.toString();
+
+                y = tbl_currentusers.getValueAt(maxIndex, 1);
+                buff_username = y.toString();
+
+                z = tbl_currentusers.getValueAt(maxIndex, 2);
+                buff_fullname = z.toString();
+
+                w = tbl_currentusers.getValueAt(maxIndex, 3);
+                buff_accounttype = w.toString();
+
+                try {
+                    int i = 0;
+
+                    dbconnect conn4 = new dbconnect();
+                    ResultSet rset1 = null;
+
+                    rset1 = conn4.getUserlist();
+
+                    while (rset1.next() && i <= maxIndex)
+                    {
+                        buff_password = rset1.getString("password");
+                        i++;
+                    }
+
+                    conn4.close();
+                } catch (SQLException throwables) {
+                    throwables.printStackTrace();
+                } catch (Exception exception) {
+                    exception.printStackTrace();
+                }
+
+                System.out.println(buff_password);
                 btn_delete.setEnabled(true);
                 btn_modify.setEnabled(true);
             }
